@@ -34,8 +34,8 @@
         "leaflet-search": "^3.0.9",
         "primeicons": "^6.0.1",
         "primereact": "^8.7.3",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.1.0",
+        "react-dom": "^18.1.0",
         "react-google-maps": "^9.4.5",
         "react-helmet": "^6.1.0",
         "react-leaflet": "^4.2.0",
@@ -44,6 +44,7 @@
         "react-slick": "^0.29.0",
         "react-star-ratings": "^2.3.0",
         "react-verification-code-input": "^1.2.9",
+        "socket.io-client": "^4.6.1",
         "sweetalert2": "^11.6.16",
         "sweetalert2-react-content": "^5.0.7",
         "swiper": "^8.4.5",
