@@ -4,16 +4,17 @@ const { Jsonwt } = require("../helpers/jwt");
 
 const userStore = async(req, res) => {
     const { email, password } = req.body;
-
     try {
         const searchEmail = await User.findOne({ where: { email } })
+
         if(searchEmail){
             return res.status(409).json({ msg: 'El correo ingresado ya existe' })
         }
-
+        
         const hashedPassword = await bcrypt.hash(password, 10);
-    
+        
         const user = await User.create({ email, password: hashedPassword })
+
         if(!user){
             return res.status(500).json({ msg: 'Hubo un problema, intente nuevamente' })
         }
@@ -33,14 +34,12 @@ const userLogin = async(req, res) => {
     
     try {
         //Verificar si el email existe
-        const user = await User.findOne({ email })
-        console.log(user)
+        const user = await User.findOne({ where: { email } })
+
         if(!user){
             return res.status(400).json({ msg: "El correo o contraseña son incorrectos" })
         }
-        if(user.email !== email){
-            return res.status(400).json({ msg: "El correo o contraseña son incorrectos" })
-        }
+
         //Verificar la contraseña
         const validatePassword = bcrypt.compareSync(password, user.password)
         if(!validatePassword){
