@@ -3,7 +3,7 @@ const { sequelize } = require("../database");
 
 const User = sequelize.define('user', {
     id: {
-      type: DataTypes.UUID,
+      type: DataTypes.STRING,
       defaultValue: DataTypes.UUIDV4,
       allowNull: false,
       primaryKey: true
