@@ -3,10 +3,10 @@ import Footer from "../Footer/Footer";
 import Header from "../Header/Header";
 
 export default function ({ children }) {
-   
+    const auth = localStorage.getItem('auth')
     return (
         <div className="Guest">
-            <Header />
+            <Header auth={auth}/>
                 {children}
             <Footer />
         </div>
