@@ -9,32 +9,6 @@ import ansiedad from "../../../assets/guest/ansiedad.png";
 import mentalhealth from "../../../assets/guest/mentalhealth.png"
 import logoSinFondo from "../../../assets/guest/logoSinFondo.png"
 
-import Bed from "../../../assets//icons/test/034-bed.svg";
-import Vela from "../../../assets//icons/test/Combined Shape.svg";
-import House from "../../../assets//icons/test/Combined Shape (1).svg";
-import Botella from "../../../assets//icons/test/Combined Shape (2).svg";
-import Granja from "../../../assets//icons/test/Combined Shape (3).svg";
-import Food from "../../../assets//icons/test/Combined Shape (4).svg";
-import Autencilios from "../../../assets//icons/test/Combined Shape (5).svg";
-import Sol from "../../../assets//icons/test/Combined Shape (6).svg";
-import Globo from "../../../assets//icons/test/globo-aerostatico.svg";
-
-import ExpA from "../../../assets/icons/expA.svg";
-import ExpB from "../../../assets/icons/expB.svg";
-import ExpC from "../../../assets/icons/expC.svg";
-import ExpD from "../../../assets/icons/expD.svg";
-import ExpE from "../../../assets/icons/expE.svg";
-import ExpF from "../../../assets/icons/expF.svg";
-
-import GasA from "../../../assets/icons/gasA.svg";
-import GasB from "../../../assets/icons/gasB.svg";
-import GasC from "../../../assets/icons/gasC.svg";
-import GasD from "../../../assets/icons/gasD.svg";
-import GasE from "../../../assets/icons/gasE.svg";
-import GasF from "../../../assets/icons/gasF.svg";
-
-import Experiencias from "../../../assets/icons/Experiencias.svg";
-
 import "./LandingPage.scss";
 import Carousel from "../../../components/molecules/Carousel/Carousel";
 import { useNavigate } from "react-router-dom";
@@ -51,6 +25,7 @@ import alimentos from "../../../assets/guest/alimentos.png";
 const LandingPage = () => {
     const [openPlace, setOpenPlace] = useState(false);
     const [openCategory, setOpenCategory] = useState(false); //-250
+    const auth = localStorage.getItem('auth')
 
     useEffect(() => {
         if (typeof window !== "undefined") {
@@ -91,139 +66,6 @@ const LandingPage = () => {
         },
     ];
 
-    const [destPlace, setDestPlace] = useState([
-        {
-            id: 1,
-            place: "Aguascalientes",
-        },
-        {
-            id: 2,
-            place: "Baja california",
-        },
-        {
-            id: 3,
-            place: "Chihuahua",
-        },
-        {
-            id: 4,
-            place: "Durango",
-        },
-        {
-            id: 5,
-            place: "Guanajato",
-        },
-        {
-            id: 6,
-            place: "Jalisco",
-        },
-        {
-            id: 7,
-            place: "Puebla",
-        },
-        {
-            id: 8,
-            place: "Queretaro",
-        },
-        {
-            id: 9,
-            place: "Zacatecas",
-        },
-        {
-            id: 10,
-            place: "Sonora",
-        },
-        {
-            id: 11,
-            place: "Coahuila",
-        },
-        {
-            id: 12,
-            place: "Nuevo León",
-        },
-        {
-            id: 13,
-            place: "San Luis Potosi",
-        },
-    ]);
-
-    const [category, setCategory] = useState([
-        {
-            id: 1,
-            category: "Avión",
-        },
-        {
-            id: 2,
-            category: "Motocicleta",
-        },
-        {
-            id: 3,
-            category: "Museos",
-        },
-        {
-            id: 4,
-            category: "Globo",
-        },
-        {
-            id: 5,
-            category: "Tren",
-        },
-        {
-            id: 6,
-            category: "Avión",
-        },
-        {
-            id: 7,
-            category: "Motocicleta",
-        },
-        {
-            id: 8,
-            category: "Museos",
-        },
-        {
-            id: 9,
-            category: "Globo",
-        },
-    ]);
-
-    const anfintriones = [
-        {
-            img: House,
-            title: "Viñedos",
-        },
-        {
-            img: Sol,
-            title: "Naturaleza",
-        },
-        {
-            img: Vela,
-            title: "Welness",
-        },
-        {
-            img: Bed,
-            title: "Hospedaje",
-        },
-        {
-            img: Food,
-            title: "Cartas de vino",
-        },
-        {
-            img: Botella,
-            title: "Prodductos",
-        },
-        {
-            img: Globo,
-            title: "Experiencias",
-        },
-        {
-            img: Granja,
-            title: "Eno-Tours",
-        },
-        {
-            img: Autencilios,
-            title: "Gastronomia",
-        },
-    ];
-
     const blog = [
         {
             id: 1,
@@ -244,39 +86,6 @@ const LandingPage = () => {
             descriptions: "Comer descontroladamente, hacer dietas rigurosas, provocarse vómitos, abuso de laxantes y diuréticos, problemas con la dentadura, depresiones detectables, ir al baño después de comer, menstruaciones irregulares y preocupación exagerada por el peso.",
         },
         
-    ];
-
-    const products = [
-        {
-            note: true,
-            colab: true,
-            image: Experiencias,
-        },
-        {
-            note: false,
-            colab: false,
-            image: Experiencias,
-        },
-        {
-            note: true,
-            colab: true,
-            image: Experiencias,
-        },
-        {
-            note: false,
-            colab: false,
-            image: Experiencias,
-        },
-        {
-            note: true,
-            colab: true,
-            image: Experiencias,
-        },
-        {
-            note: false,
-            colab: false,
-            image: Experiencias,
-        },
     ];
 
     const vinedo = [
@@ -312,151 +121,6 @@ const LandingPage = () => {
             id: 6,
             img: c7,
             
-        },
-    ];
-
-    const exper = [
-        {
-            id: 1,
-            img: ExpA,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-        {
-            id: 2,
-            img: ExpB,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-        {
-            id: 3,
-            img: ExpC,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-        {
-            id: 4,
-            img: ExpD,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-        {
-            id: 5,
-            img: ExpE,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-        {
-            id: 6,
-            img: ExpF,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-    ];
-
-    const gastro = [
-        {
-            id: 1,
-            img: GasA,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-        {
-            id: 2,
-            img: GasB,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-        {
-            id: 3,
-            img: GasC,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-        {
-            id: 4,
-            img: GasD,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-        {
-            id: 5,
-            img: GasE,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-        {
-            id: 6,
-            img: GasF,
-            price: "$500.00",
-            priceReal: "$700.00",
-            title: "Paseo matutino",
-            number: "(1512)"
-        },
-    ];
-
-    const ubicationPlace = [
-        {
-            lat: 30.153815,
-            lng: -115.23181,
-            color: "rgba(242, 174, 212, 0.7)",
-        },
-        {
-            lat: 29.899816,
-            lng: -111.761276,
-            color: "rgba(255, 200, 68, 0.7)",
-        },
-        {
-            lat: 30.947496,
-            lng: -109.417839,
-            color: "rgba(180, 189, 0, 0.7)",
-        },
-        {
-            lat: 26.774068,
-            lng: -107.976654,
-            color: "rgba(125, 185, 245, 0.7)",
-        },
-        {
-            lat: 25.896449,
-            lng: -98.62743,
-            color: "rgba(226, 164, 252, 0.7)",
-        },
-        {
-            lat: 16.845153,
-            lng: -96.765453,
-            color: "rgba(242, 174, 212, 0.7)",
-        },
-        {
-            lat: 20.632944,
-            lng: -88.647916,
-            color: "rgba(255, 200, 68, 0.7)",
-        },
-        {
-            lat: 20.6597712,
-            lng: -103.349762,
-            color: "rgba(255, 0, 0, 0.7)",
         },
     ];
 
@@ -562,14 +226,12 @@ const LandingPage = () => {
                     </div>
                     
                 </div>
+                {!auth ?
                 <div className="secct7">
                     <h1 className="secct7-title">¿Quieres realizar el test?</h1>
                     <div className="secct7-bodyA">
                     Realiza el test para saber tu posible diagnóstico, si tienes algún problema que afecte tu salud mental.
                     </div>
-                    <p className="secct7-bodyA marg">
-                        Si quieres saber más, da click aqui
-                    </p>
 
                     <div className="secct7-btn">
                         <div className="secct-7-btn">
@@ -581,8 +243,17 @@ const LandingPage = () => {
                         </div>
                     </div>
                 </div>
+                :
+                <div className="secct7">
+                    <h1 className="secct7-title">Better U</h1>
+                    <div className="secct7-bodyA">
+                        Cuando dicen "vendrá algo mejor", no siempre significa otra persona. También es estabilidad
+                        contigo mismo, libertad, madurez emocional, amor propio y, sobre todo, paz mental.
+                    </div>
+                </div>
+                }
                 <div className="secct8">
-                    <h1 className="secct8-title">Para saber màs...</h1>
+                    <h1 className="secct8-title">Para saber más...</h1>
                     <div className="secct-8-body">
                         {blog.map((blog) => (
                             <div className="card-blog" key={blog.id}>
