@@ -1,22 +1,10 @@
 import "./Test.scss";
 
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import ChatAlkimia from "../../../components/molecules/ChatAlkimia/ChatAlkimia";
 
 const Test = () => {
 
-    const [mensajes, setMensajes] = useState([
-        {
-            autor: 'Cliente',
-            contenido: 'Hola, buen día. Tengo problemas con mi compra',
-            tiempo: new Date()
-        },
-        {
-            autor: 'Proveedor',
-            contenido: 'Hola, Santiago ¿Cómo podemos ayudarte?',
-            tiempo: new Date()
-        },
-    ]);
     return (
         <div className="AuthView">
             <div className="Contact-Better">
@@ -25,7 +13,7 @@ const Test = () => {
                 </div>
                 <div className="cnt-txt">Test Better U</div>
                 <div className="cardBody-Better">
-                    <ChatAlkimia mensajes={mensajes} setMensajes={setMensajes}/>
+                    <ChatAlkimia />
                 </div>
             </div>
         </div>
